pyproject.toml
[project]
name = "rag-dynamic"
version = "0.1.0"
description = "Dynamic Agentic RAG with File Uploads"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "pypdf>=5.1.0",
    "openai>=2.14.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
    "wikipedia>=1.4.0",
]